refactor(discord): extract task response fence store
